+++ braintree/transaction_amounts.py
@@ -2,6 +2,7 @@ class TransactionAmounts(object):
     """ A class of constants for transaction amounts that will cause different statuses. """
 
     Authorize = "1000.00"
+    PartiallyAuthorized = "1004.00"
     Decline = "2000.00"
     HardDecline = "2015.00"
     Fail = "3000.00"